+++ agents/rust/aries-vcx-agent/src/agent/agent_config.rs
@@ -1,5 +1,5 @@
 use aries_vcx::agency_client::configuration::AgencyClientConfig;
-use aries_vcx_core::wallet::indy_wallet::{IssuerConfig, WalletConfig};
+use aries_vcx_core::wallet::indy::{IssuerConfig, WalletConfig};
 
 #[derive(Clone)]
 pub struct AgentConfig {
