+++ agents/rust/aries-vcx-agent/src/agent/init.rs
@@ -11,7 +11,7 @@ use aries_vcx_core::indy::{
     wallet::{create_wallet_with_master_secret, open_wallet, wallet_configure_issuer},
 };
 use url::Url;
-use aries_vcx_core::wallet::indy_wallet::WalletConfig;
+use aries_vcx_core::wallet::indy::WalletConfig;
 
 use crate::{
     agent::{agent_config::AgentConfig, agent_struct::Agent},
