+++ aries_vcx/tests/utils/devsetup_util.rs
@@ -12,7 +12,7 @@ use aries_vcx_core::indy::wallet::{create_wallet_with_master_secret, open_wallet
 #[cfg(feature = "modular_libs")]
 use aries_vcx_core::ledger::request_submitter::vdr_ledger::LedgerPoolConfig;
 use aries_vcx_core::wallet::base_wallet::BaseWallet;
-use aries_vcx_core::wallet::indy_wallet::{IndySdkWallet, WalletConfig};
+use aries_vcx_core::wallet::indy::{IndySdkWallet, WalletConfig};
 use aries_vcx_core::WalletHandle;
 
 use crate::utils::devsetup_alice::Alice;
@@ -29,7 +29,7 @@ pub mod test_utils {
     use aries_vcx_core::indy::wallet::{close_wallet, delete_wallet};
     #[cfg(feature = "modular_libs")]
     use aries_vcx_core::ledger::request_submitter::vdr_ledger::LedgerPoolConfig;
-    use aries_vcx_core::wallet::indy_wallet::WalletConfig;
+    use aries_vcx_core::wallet::indy::WalletConfig;
     use aries_vcx_core::WalletHandle;
     use messages::msg_fields::protocols::connection::Connection;
     use messages::msg_fields::protocols::cred_issuance::CredentialIssuance;
