+++ aries_vcx_core/src/indy/wallet_non_secrets.rs
@@ -1,7 +1,7 @@
 use serde_json;
 
 use crate::errors::error::{AriesVcxCoreError, AriesVcxCoreErrorKind, VcxCoreResult};
-use crate::wallet::internal::{add_wallet_record, delete_wallet_record, get_wallet_record, update_wallet_record_value};
+use crate::wallet::indy::internal::{add_wallet_record, delete_wallet_record, get_wallet_record, update_wallet_record_value};
 use crate::WalletHandle;
 
 static WALLET_RECORD_TYPE: &str = "cache";
