+++ aries_vcx_core/src/wallet/indy_wallet.rs
@@ -1,244 +1,1 @@
-use std::collections::HashMap;
-use std::thread;
 
-use async_trait::async_trait;
-use futures::executor::block_on;
-use serde_json::Value;
-use serde::{Deserialize, Serialize};
-
-use crate::errors::error::{AriesVcxCoreError, AriesVcxCoreErrorKind, VcxCoreResult};
-use crate::{indy, utils::{async_fn_iterator::AsyncFnIterator, json::TryGetIndex}, wallet};
-use crate::{SearchHandle, WalletHandle};
-
-use super::base_wallet::BaseWallet;
-
-#[derive(Debug)]
-pub struct IndySdkWallet {
-    pub wallet_handle: WalletHandle,
-}
-
-impl IndySdkWallet {
-    pub fn new(wallet_handle: WalletHandle) -> Self {
-        IndySdkWallet { wallet_handle }
-    }
-}
-
-#[allow(unused_variables)]
-#[async_trait]
-impl BaseWallet for IndySdkWallet {
-    async fn create_and_store_my_did(
-        &self,
-        seed: Option<&str>,
-        method_name: Option<&str>,
-    ) -> VcxCoreResult<(String, String)> {
-        indy::wallet::create_and_store_my_did(self.wallet_handle, seed, method_name).await
-    }
-
-    async fn key_for_local_did(&self, did: &str) -> VcxCoreResult<String> {
-        indy::wallet::get_verkey_from_wallet(self.wallet_handle, did).await
-    }
-
-    async fn replace_did_keys_start(&self, target_did: &str) -> VcxCoreResult<String> {
-        indy::wallet::libindy_replace_keys_start(self.wallet_handle, target_did).await
-    }
-
-    async fn replace_did_keys_apply(&self, target_did: &str) -> VcxCoreResult<()> {
-        indy::wallet::libindy_replace_keys_apply(self.wallet_handle, target_did).await
-    }
-
-    async fn add_wallet_record(
-        &self,
-        xtype: &str,
-        id: &str,
-        value: &str,
-        tags: Option<HashMap<String, String>>,
-    ) -> VcxCoreResult<()> {
-        let res = tags.map(|x| serde_json::to_string(&x)).transpose()?.to_owned();
-        let tags_json = res.as_deref();
-        wallet::internal::add_wallet_record(self.wallet_handle, xtype, id, value, tags_json).await
-    }
-
-    async fn get_wallet_record(&self, xtype: &str, id: &str, options: &str) -> VcxCoreResult<String> {
-        wallet::internal::get_wallet_record(self.wallet_handle, xtype, id, options).await
-    }
-
-    async fn get_wallet_record_value(&self, xtype: &str, id: &str) -> VcxCoreResult<String> {
-        let options = r#"{"retrieveType": false, "retrieveValue": true, "retrieveTags": false}"#;
-
-        let str_record = self.get_wallet_record(xtype, id, options).await?;
-        let wallet_record: WalletRecord = serde_json::from_str(&str_record)?;
-        wallet_record.value.ok_or_else(|| {
-            AriesVcxCoreError::from_msg(
-                AriesVcxCoreErrorKind::WalletRecordNotFound,
-                "The wallet record does not have a value",
-            )
-        })
-    }
-
-    async fn delete_wallet_record(&self, xtype: &str, id: &str) -> VcxCoreResult<()> {
-        wallet::internal::delete_wallet_record(self.wallet_handle, xtype, id).await
-    }
-
-    async fn update_wallet_record_value(&self, xtype: &str, id: &str, value: &str) -> VcxCoreResult<()> {
-        wallet::internal::update_wallet_record_value(self.wallet_handle, xtype, id, value).await
-    }
-
-    async fn update_wallet_record_tags(
-        &self,
-        xtype: &str,
-        id: &str,
-        tags: HashMap<String, String>,
-    ) -> VcxCoreResult<()> {
-        let tags_json = serde_json::to_string(&tags)?;
-        wallet::internal::update_wallet_record_tags(self.wallet_handle, xtype, id, &tags_json).await
-    }
-
-    async fn add_wallet_record_tags(&self, xtype: &str, id: &str, tags: HashMap<String, String>) -> VcxCoreResult<()> {
-        let tags_json = serde_json::to_string(&tags)?;
-        wallet::internal::add_wallet_record_tags(self.wallet_handle, xtype, id, &tags_json).await
-    }
-
-    async fn delete_wallet_record_tags(&self, xtype: &str, id: &str, tag_names: &str) -> VcxCoreResult<()> {
-        wallet::internal::delete_wallet_record_tags(self.wallet_handle, xtype, id, tag_names).await
-    }
-
-    async fn iterate_wallet_records(
-        &self,
-        xtype: &str,
-        query: &str,
-        options: &str,
-    ) -> VcxCoreResult<Box<dyn AsyncFnIterator<Item = VcxCoreResult<String>>>> {
-        let search = wallet::internal::open_search_wallet(self.wallet_handle, xtype, query, options).await?;
-        let iter = IndyWalletRecordIterator::new(self.wallet_handle, search);
-
-        Ok(Box::new(iter))
-    }
-
-    async fn sign(&self, my_vk: &str, msg: &[u8]) -> VcxCoreResult<Vec<u8>> {
-        indy::signing::sign(self.wallet_handle, my_vk, msg).await
-    }
-
-    async fn verify(&self, vk: &str, msg: &[u8], signature: &[u8]) -> VcxCoreResult<bool> {
-        indy::signing::verify(vk, msg, signature).await
-    }
-
-    async fn pack_message(&self, sender_vk: Option<&str>, receiver_keys: &str, msg: &[u8]) -> VcxCoreResult<Vec<u8>> {
-        indy::signing::pack_message(self.wallet_handle, sender_vk, receiver_keys, msg).await
-    }
-
-    async fn unpack_message(&self, msg: &[u8]) -> VcxCoreResult<Vec<u8>> {
-        indy::signing::unpack_message(self.wallet_handle, msg).await
-    }
-
-    fn get_wallet_handle(&self) -> WalletHandle {
-        self.wallet_handle
-    }
-}
-
-struct IndyWalletRecordIterator {
-    wallet_handle: WalletHandle,
-    search_handle: SearchHandle,
-}
-
-impl IndyWalletRecordIterator {
-    fn new(wallet_handle: WalletHandle, search_handle: SearchHandle) -> Self {
-        IndyWalletRecordIterator {
-            wallet_handle,
-            search_handle,
-        }
-    }
-
-    async fn fetch_next_records(&self) -> VcxCoreResult<Option<String>> {
-        let indy_res_json = wallet::internal::fetch_next_records_wallet(self.wallet_handle, self.search_handle, 1).await?;
-
-        let indy_res: Value = serde_json::from_str(&indy_res_json)?;
-
-        let records = (&indy_res).try_get("records")?;
-
-        let item: Option<VcxCoreResult<String>> = records
-            .as_array()
-            .and_then(|arr| arr.first())
-            .map(|item| serde_json::to_string(item).map_err(AriesVcxCoreError::from));
-
-        item.transpose()
-    }
-}
-
-/// Implementation of a generic [AsyncFnIterator] iterator for indy/vdrtools wallet record iteration.
-/// Wraps over the vdrtools record [SearchHandle] functionality
-#[async_trait]
-impl AsyncFnIterator for IndyWalletRecordIterator {
-    type Item = VcxCoreResult<String>;
-
-    async fn next(&mut self) -> Option<Self::Item> {
-        let records = self.fetch_next_records().await;
-        records.transpose()
-    }
-}
-
-impl Drop for IndyWalletRecordIterator {
-    fn drop(&mut self) {
-        let search_handle = self.search_handle;
-
-        thread::spawn(move || {
-            block_on(async {
-                wallet::internal::close_search_wallet(search_handle).await.ok();
-            });
-        });
-    }
-}
-
-#[derive(Clone, Debug, Default, Builder, Serialize, Deserialize)]
-#[builder(setter(into, strip_option), default)]
-pub struct WalletConfig {
-    pub wallet_name: String,
-    pub wallet_key: String,
-    pub wallet_key_derivation: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub wallet_type: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub storage_config: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub storage_credentials: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub rekey: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub rekey_derivation_method: Option<String>,
-}
-
-#[derive(Clone, Debug, Default, Builder, Serialize, Deserialize)]
-#[builder(setter(into, strip_option), default)]
-pub struct IssuerConfig {
-    pub institution_did: String,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-struct WalletCredentials {
-    key: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    rekey: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    storage_credentials: Option<serde_json::Value>,
-    key_derivation_method: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    rekey_derivation_method: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct WalletRecord {
-    id: Option<String>,
-    #[serde(rename = "type")]
-    record_type: Option<String>,
-    pub value: Option<String>,
-    tags: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct RestoreWalletConfigs {
-    pub wallet_name: String,
-    pub wallet_key: String,
-    pub exported_wallet_path: String,
-    pub backup_key: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub wallet_key_derivation: Option<String>,
-}
