+++ aries_vcx_core/src/wallet/mod.rs
@@ -3,3 +3,4 @@ pub mod base_wallet;
 #[cfg(feature = "vdrtools")]
 pub mod indy_wallet;
 pub mod internal;
+pub mod indy;
